+++ Develappers.RedmineHourglassApi/HttpClient.cs
@@ -4,6 +4,7 @@ using System.Net;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Develappers.RedmineHourglassApi.Logging;
 
 namespace Develappers.RedmineHourglassApi
 {
@@ -24,35 +25,29 @@ namespace Develappers.RedmineHourglassApi
             _apiKey = apiKey;
         }
 
-        public Task<string> DeleteAsync(Uri relativeUri, CancellationToken token = default(CancellationToken))
-        {
-            return ExecuteRequestInternalAsync("DELETE", relativeUri, null, token);
-        }
-
-        public Task<string> GetStringAsync(Uri relativeUri, CancellationToken token = default(CancellationToken))
-        {
-            return ExecuteRequestInternalAsync("GET", relativeUri, null, token);
-        }
-
         private async Task<string> ExecuteRequestInternalAsync(string method, Uri relativeUri, string value, CancellationToken token)
         {
             if (string.IsNullOrEmpty(method))
             {
+                LogProvider.GetCurrentClassLogger().Error($"{nameof(ExecuteRequestInternalAsync)} called with invalid method");
                 throw new ArgumentNullException(nameof(method));
             }
 
             if (method != "GET" && method != "PUT" && method != "POST" && method != "DELETE")
             {
+                LogProvider.GetCurrentClassLogger().Error($"{nameof(ExecuteRequestInternalAsync)} called with unsupported method");
                 throw new ArgumentException("invalid request method", nameof(method));
             }
 
             if (relativeUri == null)
             {
-                throw new ArgumentNullException(nameof(method));
+                throw new ArgumentNullException(nameof(relativeUri));
             }
 
             var baseUri = new Uri(_hourglassUrl);
             var completeUri = new Uri(baseUri, relativeUri);
+
+            LogProvider.GetCurrentClassLogger().Debug($"executing {method} request to {completeUri} with body '{value}'" );
 
             var httpWebRequest = (HttpWebRequest)WebRequest.Create(completeUri);
             httpWebRequest.Method = method;
@@ -86,7 +81,13 @@ namespace Develappers.RedmineHourglassApi
                 }
             }
 
+            LogProvider.GetCurrentClassLogger().Debug($"retrieved result '{result}'");
             return result;
+        }
+
+        public Task<string> GetStringAsync(Uri relativeUri, CancellationToken token = default(CancellationToken))
+        {
+            return ExecuteRequestInternalAsync("GET", relativeUri, null, token);
         }
 
         public Task<string> PutStringAsync(Uri relativeUri, string value, CancellationToken token)
@@ -97,6 +98,11 @@ namespace Develappers.RedmineHourglassApi
         public Task<string> PostStringAsync(Uri relativeUri, string value, CancellationToken token)
         {
             return ExecuteRequestInternalAsync("POST", relativeUri, value, token);
+        }
+
+        public Task<string> DeleteAsync(Uri relativeUri, CancellationToken token = default(CancellationToken))
+        {
+            return ExecuteRequestInternalAsync("DELETE", relativeUri, null, token);
         }
     }
 }