+++ Develappers.RedmineHourglassApi/Types/TimeLogBookRequest.cs
@@ -5,7 +5,7 @@ namespace Develappers.RedmineHourglassApi.Types
     /// <summary>
     /// Class which wraps a <see cref="TimeBookingUpdate"/> object to be API conform.
     /// </summary>
-    internal class TimeBookingUpdateRequest
+    internal class TimeLogBookRequest
     {
         [JsonProperty("time_booking")]
         public TimeBookingUpdate Values { get; set; }