+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Contribution guidelines, code of conduct and changelog
-- Possibility to start and stop time trackers
-- Update on time bookings.
+- time trackers: start, stop, delete
+- time bookings: Update
+- support for logging frameworks
 
 ### Changed
-- GetById throws NotFoundException instead of returning null when item was not found 
+- renamed all service functions because "ById" is implicit clear
+- Get throws NotFoundException instead of returning null when item was not found 
 
 ### Removed
 - nothing